pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "alloydb-survey-mcp-server"
version = "1.0.0"
description = "MCP server for fetching on-ground survey data from Alloy DB"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "mcp>=1.0.0",
    "google-cloud-alloydb-connectors>=1.0.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
]

[project.scripts]
alloydb-survey-server = "alloydb_survey_mcp.server:main"
